new font, emoji, cookie and other

+++ game/views/reset.py
@@ -9,39 +9,41 @@ from game.models.BusinessPayments import BusinessPayments
 
 def reset(request):
 
-    # Delete all about last game
-    Player.objects.all().delete()
-    Actions.objects.all().delete()
-    PlayersBusiness.objects.all().delete()
-    CommandPayments.objects.all().delete()
-    BusinessPayments.objects.all().delete()
-
-    player_X_balance = 200000
-
-    player_X = Player( name="X", visible=False )
-    player_X.save()
+    if request.user.is_authenticated:
     
-    commandBusinessPayments = CommandPayments(
-        player = player_X, 
-        count  = player_X_balance,
-    )
-    commandBusinessPayments.save()
+        # Delete all about last game
+        Player.objects.all().delete()
+        Actions.objects.all().delete()
+        PlayersBusiness.objects.all().delete()
+        CommandPayments.objects.all().delete()
+        BusinessPayments.objects.all().delete()
 
-    name = f'''Начал игру - {player_X_balance}'''
-    Actions(
-        player   = player_X,
-        name     = name,
-        count    = player_X_balance,
-        category = 'OTHER',
-    ).save()
+        player_X_balance = 200000
 
-    name = f'''Вложил в командный бизнес - {player_X_balance}'''
-    
-    Actions(
-        player   = player_X,
-        name     = name,
-        count    = -player_X_balance,
-        category = 'CMND',
-    ).save()
+        player_X = Player( name="X", visible=False )
+        player_X.save()
+        
+        commandBusinessPayments = CommandPayments(
+            player = player_X, 
+            count  = player_X_balance,
+        )
+        commandBusinessPayments.save()
+
+        name = f'''Начал игру - {player_X_balance}'''
+        Actions(
+            player   = player_X,
+            name     = name,
+            count    = player_X_balance,
+            category = 'OTHER',
+        ).save()
+
+        name = f'''Вложил в командный бизнес - {player_X_balance}'''
+        
+        Actions(
+            player   = player_X,
+            name     = name,
+            count    = -player_X_balance,
+            category = 'CMND',
+        ).save()
 
     return redirect("/")
